requirements.txt
django
psycopg2-binary
celery==4.4.7
djangorestframework
django-celery-results
Pillow
selenium
beautifulsoup4
html5lib
googletrans
pydrive
pandas
redis
gunicorn
django-heroku
whitenoise
celery-progress
tesserocr

